snet-service/requirements.txt
grpcio==1.27.2
grpcio-tools==1.27.2
requests==2.20.0
